More WIP on RNG validation:
 - added uri() method to file storage to retreive the (potentially
   time-limited) URI of a file given its classifier (bucket) and
   path (key)
 - moved most of the 'tools' admin endpoints to a Tools controller
 - added various experimental ways to validate an EAD, such as via
   a POST request or via a Cloud-stored file descriptor
