+++ build.sbt
@@ -102,7 +102,11 @@ val portalDependencies = Seq(
   "com.lightbend.akka" %% "akka-stream-alpakka-s3" % "1.1.2",
 
   // EAD validation testing
-  "org.relaxng" % "jing" % "20181222"
+  "org.relaxng" % "jing" % "20181222",
+
+  // S3 sdk
+  "com.amazonaws" % "aws-java-sdk-core" % "1.11.476",
+  "com.amazonaws" % "aws-java-sdk-s3" % "1.11.476"
 )
 
 val testDependencies = Seq(
