+++ modules/admin/test/resources/invalid-ead.xml
@@ -1,0 +1,111 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!-- <!DOCTYPE ead PUBLIC "+//ISBN 1-931666-00-8//DTD ead.dtd (Encoded Archival Description (EAD) Version 2002)//EN" "http://lcweb2.loc.gov/xmlcommon/dtds/ead2002/ead.dtd"> -->
+
+<ead xmlns="urn:isbn:1-931666-22-9">
+	<eadheader langencoding="iso639-2b" countryencoding="iso3166-1"
+		dateencoding="iso8601" repositoryencoding="iso15511" scriptencoding="iso15924"
+		relatedencoding="DC">
+		<eadid countrycode="GB" mainagencycode="GB00001" encodinganalog="Identifier">test_desc_id</eadid>
+		<filedesc>
+			<titlestmt>
+				<titleproper encodinganalog="Title">Test EAD Item</titleproper>
+			</titlestmt>
+			<publicationstmt>
+				<publisher encodinganalog="Publisher">Test Repository</publisher>
+				<address>
+					<addressline>Test Address 1</addressline>
+					<addressline>Test Address 2</addressline>
+				</address>
+				<date normal="20120610" encodinganalog="Date">2012-06-10</date>
+			</publicationstmt>
+		</filedesc>
+		<profiledesc>
+			<creation>
+				Generated by qubit 1.2
+				<date normal="20120925">2012-09-25 14:44:UTC</date>
+			</creation>
+			<langusage>
+				<language langcode="eng" encodinganalog="Language">English</language>
+			</langusage>
+			<descrules encodinganalog="3.7.2">ISAD(G)</descrules>
+		</profiledesc>
+	</eadheader>
+	<!-- TODO <frontmatter/> -->
+	<archdesc level="fonds" relatedencoding="ISAD(G)v2">
+            <did>
+					<unittitle encodinganalog="3.1.2">Test EAD Item</unittitle>
+					<unitid repositorycode="GB-00001" countrycode="GB"
+						encodinganalog="3.1.1">Ctop level fonds</unitid>
+                
+            </did>
+		<dsc type="combined">
+			<c01 level="subfonds">
+				<did>
+					<unittitle encodinganalog="3.1.2">Test EAD Item</unittitle>
+					<unitid repositorycode="GB-00001" countrycode="GB"
+						encodinganalog="3.1.1">C00001</unitid>
+					<unitdate datechar="creation" normal="19240101/19471231"
+						encodinganalog="3.1.3">1924-1-1 - 1947-12-31</unitdate>
+					<origination encodinganalog="3.2.1">
+						<name>Test Name</name>
+					</origination>
+					<physdesc>
+						<extent encodinganalog="3.1.5">167 files</extent>
+					</physdesc>
+					<repository>
+						<corpname>Test Corportate Body</corpname>
+						<address>
+							<addressline>Test Address 1</addressline>
+							<addressline>Test Address 2</addressline>
+						</address>
+					</repository>
+					<langmaterial encodinganalog="3.4.3">
+						<language langcode="eng">English</language>
+						<language langcode="fre">French</language>
+					</langmaterial>
+				</did>
+				<scopecontent encodinganalog="3.3.1">
+					<p>
+						Top level item.
+					</p>
+				</scopecontent>
+				<c02>
+					<did>
+						<unittitle encodinganalog="3.1.2">Test EAD Item Nested 1</unittitle>
+						<unitid repositorycode="GB-00001" countrycode="GB"
+							encodinganalog="3.1.1">C00002</unitid>
+					</did>
+					<scopecontent encodinganalog="3.3.1">
+						<p>
+							First nested item.
+						</p>
+					</scopecontent>
+					<c03>
+						<did>
+							<unittitle encodinganalog="3.1.2">Test EAD Item Nested 2 - 1</unittitle>
+							<unitid repositorycode="GB-00001" countrycode="GB"
+								encodinganalog="3.1.1">C00002-1</unitid>
+						</did>
+						<scopecontent encodinganalog="3.3.1">
+							<p>
+								Second nested item.
+							</p>
+						</scopecontent>
+					</c03>
+					<c03>
+						<did>
+							<unittitle encodinganalog="3.1.2">Test EAD Item Nested 2 - 2</unittitle>
+							<unitid repositorycode="GB-00001" countrycode="GB"
+								encodinganalog="3.1.1">C00002-2</unitid>
+						</did>
+						<scopecontent encodinganalog="3.3.1">
+							<p>
+								Second nested item.
+							</p>
+						</scopecontent>
+					</c03>					
+				</c02>
+			</c01>
+		</dsc>
+	</archdesc>
+</ead>
