+++ modules/admin/test/services/ingest/EadValidatorServiceSpec.scala
@@ -2,23 +2,65 @@ package services.ingest
 
 import java.nio.file.Paths
 
+import akka.actor.ActorSystem
+import akka.http.scaladsl.model.Uri
+import akka.stream.Materializer
+import akka.stream.scaladsl.{FileIO, Sink, Source}
+import akka.util.ByteString
 import com.google.common.io.Resources
 import play.api.test.PlaySpecification
 
+import scala.concurrent.ExecutionContext
+
 class EadValidatorServiceSpec extends PlaySpecification {
 
-  private implicit val ctx = scala.concurrent.ExecutionContext.global
+  private implicit val ctx: ExecutionContext = scala.concurrent.ExecutionContext.global
+  private implicit val as: ActorSystem = ActorSystem.create("test")
+  private implicit val mat: Materializer = Materializer(as)
 
   "EAD validator service" should {
-    "return errors" in {
+    "validate a file" in {
       val validator = EadValidatorService()
 
       val file = Paths.get(Resources.getResource("valid-ead.xml").toURI)
       val errs = await(validator.validateEad(file))
       errs.size must_== 0
+    }
 
-      val errs2 = await(validator.validateEad(file))
+    "validate an URL" in {
+      val validator = EadValidatorService()
+
+      val file = Uri(Resources.getResource("valid-ead.xml").toString)
+      val errs = await(validator.validateEad(file))
+      errs.size must_== 0
+    }
+
+    "validate a stream" in {
+      val validator = EadValidatorService()
+
+      val file = Resources.getResource("valid-ead.xml")
+      val src = FileIO.fromPath(Paths.get(file.toURI))
+      val errs = await(validator.validateEad(src))
+      errs.size must_== 0
+    }
+
+    "report errors from a stream" in {
+      val validator = EadValidatorService()
+
+      val file = Resources.getResource("invalid-ead.xml")
+      val src = FileIO.fromPath(Paths.get(file.toURI))
+      val errs = await(validator.validateEad(src))
+      errs.size must_== 1
+
+      // Ensure we can repeat the validation and receive the same
+      // number of errors.
+      val src2 = FileIO.fromPath(Paths.get(file.toURI))
+      val errs2 = await(validator.validateEad(src2))
       errs2.size must_== errs.size
+
+      errs2.head.line must_== 8
+      errs2.head.pos must_== 80
+      errs2.head.error must contain("value of attribute \"mainagencycode\" is invalid")
     }
   }
 }
