+++ test/services/storage/MockFileStorage.scala
@@ -9,13 +9,14 @@ import akka.stream.Materializer
 import akka.stream.scaladsl.{FileIO, Source}
 import akka.util.ByteString
 
+import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
 
 
 case class MockFileStorage(fakeFiles: collection.mutable.ListBuffer[FileStorage#File]) extends FileStorage {
 
   private implicit val as: ActorSystem = ActorSystem("test")
-  private implicit val mat = Materializer(as)
+  private implicit val mat: Materializer = Materializer(as)
   private implicit val ec: ExecutionContext = mat.executionContext
   private def urlPrefix(classifier: String) = s"https://$classifier.mystorage.com/"
 
@@ -34,4 +35,7 @@ case class MockFileStorage(fakeFiles: collection.mutable.ListBuffer[FileStorage#
   override def listFiles(classifier: String, prefix: Option[String]): Source[FileStorage#File, NotUsed] =
     Source(fakeFiles
       .filter(p => prefix.forall(p.key.startsWith)).toList)
+
+  override def uri(classifier: String, key: String, duration: FiniteDuration = 10.minutes): URI =
+    new URI(urlPrefix(classifier) + key)
 }
